fixes #62 move seldom used parts away from ConfigProvider
A power-user can use the ConfigProviderResolver directly to access the ConfigBuilder.
That way the casual user doesn't get hit with all the complexity he doesn't need most times.
