+++ api/src/main/java/org/eclipse/microprofile/config/ConfigProvider.java
@@ -20,8 +20,6 @@
 package org.eclipse.microprofile.config;
 
 import org.eclipse.microprofile.config.spi.ConfigProviderResolver;
-import org.eclipse.microprofile.config.spi.ConfigSource;
-import org.eclipse.microprofile.config.spi.Converter;
 
 /**
  * <p>
@@ -41,14 +39,14 @@ import org.eclipse.microprofile.config.spi.Converter;
  * </p>
  *
  * <p>
- * A 'Configuration' consists of the information collected from the registered {@link ConfigSource ConfigSources}.
- * These {@link ConfigSource ConfigSources} get sorted according to
- * their <em>ordinal</em> defined via {@link ConfigSource#getOrdinal()}.
+ * A 'Configuration' consists of the information collected from the registered {@link org.eclipse.microprofile.config.spi.ConfigSource ConfigSources}.
+ * These {@link org.eclipse.microprofile.config.spi.ConfigSource ConfigSources} get sorted according to
+ * their <em>ordinal</em> defined via {@link org.eclipse.microprofile.config.spi.ConfigSource#getOrdinal()}.
  * That way it is possible to overwrite configuration with lower importance from outside.
  * </p>
  *
  * <p>
- * It is also possible to register custom {@link ConfigSource ConfigSources} to flexibly
+ * It is also possible to register custom {@link org.eclipse.microprofile.config.spi.ConfigSource ConfigSources} to flexibly
  * extend the configuration mechanism. An example would be to pick up
  * configuration values from a database table.</p>
  *
@@ -73,7 +71,7 @@ public final class ConfigProvider {
     }
 
     /**
-     * Provide a {@link Config} based on all {@link ConfigSource ConfigSources} of the
+     * Provide a {@link Config} based on all {@link org.eclipse.microprofile.config.spi.ConfigSource ConfigSources} of the
      * current Thread Context ClassLoader (TCCL)
      * The {@link Config} will be stored for future retrieval.
      * <p>
@@ -85,7 +83,7 @@ public final class ConfigProvider {
     }
 
     /**
-     * Provide a {@link Config} based on all {@link ConfigSource ConfigSources} of the
+     * Provide a {@link Config} based on all {@link org.eclipse.microprofile.config.spi.ConfigSource ConfigSources} of the
      * specified ClassLoader
      *
      * <p>
@@ -96,93 +94,4 @@ public final class ConfigProvider {
         return INSTANCE.getConfig(cl);
     }
 
-    /**
-     * Create a fresh {@link ConfigBuilder} instance. This ConfigBuilder will
-     * initially contain no {@link ConfigSource} but with default {@link Converter Converters} and auto discovered
-     * {@link ConfigSource configsources} and {@link Converter converters}.
-     * Other undiscovered {@link ConfigSource} and {@link Converter Converters} will have to be added manually.
-     *
-     * The ConfigProvider will not manage the Config instance internally
-     */
-    public static ConfigBuilder getBuilder() {
-        return INSTANCE.getBuilder();
-    }
-
-    /**
-     * Register a given {@link Config} within the Application (or Module) identified by the given ClassLoader.
-     * If the ClassLoader is {@code null} then the current Application will be used.
-     *
-     * @param config
-     *          which should get registered
-     * @param classLoader
-     *          which identifies the Application or Module the given Config should get associated with.
-     *
-     * @throws IllegalStateException
-     *          if there is already a Config registered within the Application.
-     *          A user could explicitly use {@link #releaseConfig(Config)} for this case.
-     */
-    public static void setConfig(Config config, ClassLoader classLoader) {
-        INSTANCE.setConfig(config, classLoader);
-    }
-
-    /**
-     * A {@link Config} normally gets released if the Application it is associated with gets destroyed.
-     * Invoke this method if you like to destroy the Config prematurely.
-     *
-     * If the given Config is associated within an Application then it will be unregistered.
-     */
-    public static void releaseConfig(Config config) {
-        INSTANCE.releaseConfig(config);
-    }
-
-    /**
-     * Builder for manually creating an instance of a {@code Config}.
-     *
-     * @see ConfigProvider#getBuilder()
-     */
-    public interface ConfigBuilder {
-        /**
-         * Add the default config sources appearing on the builder's classpath
-         * including:
-         * <ol>
-         * <li>System properties</li>
-         * <li>Environment properties</li>
-         * <li>/META-INF/microprofile-config.properties</li>
-         * </ol>
-         *
-         * @return the ConfigBuilder with the default config sources
-         */
-        ConfigBuilder addDefaultSources();
-
-        /**
-         * Return the ConfigBuilder for a given classloader
-         *
-         * @param loader
-         * @return the ConfigureBuilder for the given classloader
-         */
-        ConfigBuilder forClassLoader(ClassLoader loader);
-
-        /**
-         * Add the specified {@link ConfigSource}.
-         *
-         * @param sources
-         * @return the ConfigBuilder with the configured sources
-         */
-        ConfigBuilder withSources(ConfigSource... sources);
-
-        /**
-         * Add the specified {@link Converter}
-         *
-         * @param converters
-         * @return the ConfigBuilder with the added converters
-         */
-        ConfigBuilder withConverters(Converter<?>... converters);
-
-        /**
-         * Build the {@link Config} object.
-         *
-         * @return the Config object
-         */
-        Config build();
-    }
 }
